MOD: separate tutorial module
